feat(jobdaoactor): Validate schema for URIs
Java class loader supports limited number of protocols and
in case of malformed URIs Java throws a MalformedURLException.
For this case it makes sense not to return an abstract
InternalServerError, but to return a BadRequest and be
a bit more verbose.

Sample exception:
scala> jarLoader.addURL(new URL("hdfs:///someurl"))
java.net.MalformedURLException: unknown protocol: hdfs
  at java.net.URL.<init>(URL.java:600)
  at java.net.URL.<init>(URL.java:490)
  at java.net.URL.<init>(URL.java:439)

Change-Id: Id829775ee1af3f15bf6f9df08ce06086e34a0320