feat(jobserver): Add HA setup for jobserver (#1241)
* feat(jobserver): Add HA setup for jobserver

With recent changes, spinning up multiple
jobserver instances is possible. This change
adds required changes to setup an HA jobserver
cluster.

The address(es) can be specified in comma
separated format in spark.jobserver.ha.masters
section.

At startup, jobserver automatically forms
a new cluster with all the masters or joins
the old cluster of slaves. While forming the
cluster for the first time, the first node
shouldn't fail to come up. If it does then the
secondary node will be hanging forever.

kill-context-on-supervisor-down is used in
client mode for killing the slaves if supervisor
is down. This was implemented because at the
time, we didn't had any good split brain resolver.
We were relying on auto-down to mark the node
as DOWN and kill-context-on-supervisor-down to
actually kill the node. Split brain will take
care of both now. Currenlty, this HA solution
partially breaks the kill-context-on-supervisor
feature because slaves react only on the state of
1 master. In client mode, if HA setup is required
then a proper split brain resolver should be used
with kill-context-on-supervisor-down disabled.

On the slave side, I had 2 options
Option#1: Use the new ha.masters section and use it to
directly to join the masters. If this section is not
defined then use the arguments pass in JVM
Option#2: From the master pass all masters in JVM
arguments

I opted for Option#2 because it is simple and as a
developer, I don't have another branch in my mind
on how master addresses reach the slaves.

Testing:
- Tested with 1 and 2 master setup with cluster +
  supervise mode
- Tested with client mode

Notes:
- If master goes down in cluster mode,
  and comes up with a new dynamic port, then H2
  database on slave will fail to connect. Not
  a problem in ZK dao implementation.
- If port is dynamic in HA mode, then on slave
  restart, it won't know the seed nodes to connect to.
  So, port hardcoding is required.

Change-Id: I0553d0e400d471a6de70f86b9e94096231f43939

* test(ha): Add test for multiple seed nodes

Change-Id: Ie6484f89a759dbacbcae3c5ba891fa5bc68c43c7