Fixes and multi-jvm tests for split brain resolver (#1264)
* refactor(sbr): Get SBR params in separate method

Handy for testing as then only main SBR class
needs to be mocked/rewritten.

* test(akka): Add multi-jvm tests for split-brain resolver

Introduce first multi-jvm tests for split-brain
resolver class. For the easiness of testing
KeepOldestTestDowningProvider is made, which
instead of downing nodes sends messages to testProbe.

Regarding usage of multi-jvm package, please be
aware that:
* First node in test plays the role of controller. If
first node is killed during the test, the test will crash.
All cluster operations as making network split are
made on the first node.
* It's very important to use barriers and make sure
that all nodes joined the cluster (not just started,
but joined the cluster and seen as UP).
* Testing network split awaitAssert should be used to be
sure that all nodes are seen as unreachable/reachable.
Failure detector needs couple of heartbeats to
mark the node unreachable
* Barriers can be made only on all nodes (no subset), therefore
should be always called from the main test thread.
* MultiNodeClusterSpec is in separate package
(akka.cluster) as only from this package it can
see additional methods of cluster class and use
ClusterReadView to check current cluster state.

Change-Id: I7d0e3ab3e7b1db8c2873eec030b7c541a7eb3abc

* fix(sbr): Joining nodes should be considered healthy by default

If the node has status Joining it should not be seen as unreachable
without UnreachableMember event.

Worst case scenario:
Two nodes in the cluster, plus some responsive joining nodes.
Second-oldest going down. Until now, the oldest node would consider
itself as down & alone, shutting itself down and the joining nodes
as well.
With joining == healthy the oldest stays alive and the cluster
intact with second-oldest removed out of the cluster and two joining
nodes allowed to join after the removal.

* fix(sbr): Down unreachable nodes even if last timeout was canceled

In current SBR implementation if several nodes become unreachable,
for each of them there will be an UnreachableTimeout scheduled, but
all nodes will be downed together when the last scheduled
UnreachableTimeout is received.

This issue fixes the problem that when the last UnreachableTimeout
was canceled and the corresponding node became reachable again,
the remaining unstableUnreachable nodes were not downed at all.

MultiJVM tests are updated to ignore any messages received before the
ping of the testProbe. As we reuse the cluster for the tests and
don't really down the nodes, these messages are part of the
cluster healing after the network split was removed (for instance
now nodes which become reachable first will try to down other
still unreachable nodes).

Change-Id: If0c8509c1c7c6b71fb3678d3771b6799c7ac03a2