release(jobserver): Ignore untracked files during release
Jobserver uses sbt-release plugin for release management.
This plugin requires that there are no untracked files
before release, but jobserver has sometimes untracked
files which are required later.

This change enables the option to ignore untracked files.

Release is important and untracked files should
be atleast seen once to decide if something needs to
be included in the release or not. To control this, a
documentation section for release will be added which
will also cover untracked files.

Change-Id: I5fb6cfc8af0b4a408842a3b44f69e8d13a249d22